dom/tests/mochitest/bugs/test_bug848088.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=848088
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 848088</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 848088 **/

function test(loopFor, setExpandoAt)
{
    var list = document.getElementsByTagName("audio");
    delete list.length;
    var shouldHaveExpando = false;
    var realLength = list.length;
    for (var i = 0; i < loopFor; ++i) {
        if (i == setExpandoAt) {
            // Add an expando that shadows.
            Object.defineProperty(list, "length", { value: "a" });
            shouldHaveExpando = true;
        }
        var hasExpando = (list.length != realLength);
        if (shouldHaveExpando != hasExpando) {
            return false;
        }
    }
    return true;
}

ok(test(200000, 100000), "Correctly detected expando on DOM list object");

 </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=848088">Mozilla Bug 848088</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
